@@ -3,11 +3,14 @@ package com.thuctap.bookmanage;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import java.util.UUID;
+
 @SpringBootApplication
 public class BookManageApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(BookManageApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(BookManageApplication.class, args);
+
+    }
 
 }
