@@ -1,4 +1,4 @@
-package com.example.demo.photos.userID;
+package com.thuctap.bookmanage.photos.userID;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
