@@ -1,6 +1,6 @@
-package com.thuctap.book.respository;
+package com.thuctap.bookmanage.repository;
 
-import com.thuctap.book.entity.ListBook;
+import com.thuctap.bookmanage.entity.ListBook;
 import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
