@@ -1,2 +1,8 @@
-package com.thuctap.bookmanage.repository;public interface EmailSender {
+package com.thuctap.bookmanage.repository;
+
+import javax.mail.MessagingException;
+import java.io.UnsupportedEncodingException;
+
+public interface EmailSender {
+    void send(String to, String email) throws IllegalAccessException, MessagingException, UnsupportedEncodingException, jakarta.mail.MessagingException;
 }
