@@ -1,6 +1,6 @@
-package com.thuctap.book.respository;
+package com.thuctap.bookmanage.repository;
 
-import com.thuctap.book.entity.User;
+import com.thuctap.bookmanage.entity.User;
 import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -10,13 +10,12 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 @Repository
 public interface UserRepository extends JpaRepository<User,Long> {
-    @Query("SELECT c FROM User c WHERE c.username=?1 AND c.password=?2")
-    User findByUsernameAndPassword(String username, String password);
+    @Query("SELECT c FROM User c WHERE c.email=?1 AND c.password=?2")
+    User findByUsernameAndPassword(String email, String password);
     @Query("SELECT c FROM User c WHERE c.email = ?1")
     User findByEmail(String email);
     @Query("SELECT c FROM User c WHERE c.id_user=?1")
     User findByID(Long id_user);
-    User findByUsername(String username);
     @Query("SELECT c FROM User c WHERE c.id_user=?1")
     User findUserById(Long id);
 
@@ -27,6 +26,5 @@ public interface UserRepository extends JpaRepository<User,Long> {
     @Modifying
     @Query("DELETE FROM User c WHERE c.id_user = ?1")
     void deleteUserByAdmin(Long id);
-
-    public User findByResetPassToken(String resetPasswordToken);
+    public User findByToken(String token);
 }
