@@ -1,2 +1,7 @@
-package com.thuctap.bookmanage.service;public class ChapterService {
+package com.thuctap.bookmanage.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ChapterService {
 }
